fix metadata class
